+++ react/components/AccordionFilterContainer.js
@@ -42,11 +42,9 @@ const AccordionFilterContainer = ({
   showClearByFilter,
   updateOnFilterSelectionOnMobile,
   priceRangeLayout,
-  clearPriceRange,
-  setClearPriceRange,
 }) => {
   const intl = useIntl()
-  const { getSettings } = useRuntime()
+  const { getSettings, setQuery } = useRuntime()
   const [openItem, setOpenItem] = useState(null)
   const handles = useCssHandles(CSS_HANDLES)
   const isLazyFacetsFetchEnabled =
@@ -178,8 +176,10 @@ const AccordionFilterContainer = ({
                 navigationType={navigationType}
                 initiallyCollapsed={initiallyCollapsed}
                 priceRangeLayout={priceRangeLayout}
-                clearPriceRange={clearPriceRange}
-                setClearPriceRange={setClearPriceRange}
+                onClearFilter={() => {
+                  setQuery({ priceRange: undefined })
+                }}
+                showClearByFilter={showClearByFilter}
               />
             )