+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Reset price slider when clearing all filters.
+
 ## [3.122.4] - 2023-06-15
 
 ### Fixed